+++ aluraflix_api/settings.py
@@ -29,7 +29,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    #
     'aluraflix_api.core',
+    'aluraflix_api.videos',
+    #
     'rest_framework',
     'rest_framework_simplejwt',
 ]
