+++ aluraflix_api/urls.py
@@ -4,4 +4,5 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('aluraflix_api.core.urls')),
+    path('', include('aluraflix_api.videos.urls')),
 ]
