+++ aluraflix_api/videos/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from aluraflix_api.core.models import Categoria, Video
+from aluraflix_api.videos.models import Categoria, Video
 
 admin.site.register(Video)
 admin.site.register(Categoria)
