+++ aluraflix_api/videos/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class VideosConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'aluraflix_api.videos'
