+++ aluraflix_api/videos/tests/api/categoria/test_categoria_auth.py
@@ -7,7 +7,7 @@ pytestmark = pytest.mark.django_db
 
 def test_list_create_no_token(client, categoria_info):
 
-    url = resolve_url('core:categoria-list-create')
+    url = resolve_url('videos:categoria-list-create')
 
     resp = client.post(url, data=categoria_info, format='json')
 
@@ -22,7 +22,7 @@ def test_list_create_invalid_token(client, categoria_info):
 
     client.credentials(HTTP_AUTHORIZATION='Bearer ss')
 
-    url = resolve_url('core:categoria-list-create')
+    url = resolve_url('videos:categoria-list-create')
 
     resp = client.post(url, data=categoria_info, format='json')
 
@@ -43,7 +43,7 @@ def test_list_create_invalid_token(client, categoria_info):
 
 def test_read_delete_update_no_token(client, categoria, categoria_info):
 
-    url = resolve_url('core:categoria-read-delete-update', categoria.id)
+    url = resolve_url('videos:categoria-read-delete-update', categoria.id)
 
     resp = client.post(url, data=categoria_info, format='json')
 
@@ -58,7 +58,7 @@ def test_read_delete_update_invalid_token(client, categoria, categoria_info):
 
     client.credentials(HTTP_AUTHORIZATION='Bearer ss')
 
-    url = resolve_url('core:categoria-read-delete-update', categoria.id)
+    url = resolve_url('videos:categoria-read-delete-update', categoria.id)
 
     resp = client.post(url, data=categoria_info, format='json')
 
@@ -79,7 +79,7 @@ def test_read_delete_update_invalid_token(client, categoria, categoria_info):
 
 def test_videos_by_categoria_no_token(client, categoria, categoria_info):
 
-    url = resolve_url('core:videos-by-categoria', categoria.id)
+    url = resolve_url('videos:videos-by-categoria', categoria.id)
 
     resp = client.post(url, data=categoria_info, format='json')
 
@@ -94,7 +94,7 @@ def test_videos_by_categoria_invalid_token(client, categoria, categoria_info):
 
     client.credentials(HTTP_AUTHORIZATION='Bearer ss')
 
-    url = resolve_url('core:videos-by-categoria', categoria.id)
+    url = resolve_url('videos:videos-by-categoria', categoria.id)
 
     resp = client.post(url, data=categoria_info, format='json')
