+++ aluraflix_api/videos/tests/api/categoria/test_categoria_list.py
@@ -2,7 +2,7 @@ import pytest
 from django.shortcuts import resolve_url
 from rest_framework import status
 
-END_POINT = 'core:categoria-list-create'
+END_POINT = 'videos:categoria-list-create'
 
 pytestmark = pytest.mark.django_db
