+++ aluraflix_api/videos/tests/api/videos/test_videos_delete.py
@@ -2,9 +2,9 @@ import pytest
 from django.shortcuts import resolve_url
 from rest_framework import status
 
-from aluraflix_api.core.models import Video
+from aluraflix_api.videos.models import Video
 
-END_POINT = 'core:videos-read-delete-update'
+END_POINT = 'videos:videos-read-delete-update'
 
 pytestmark = pytest.mark.django_db
