+++ aluraflix_api/videos/tests/api/videos/test_videos_read.py
@@ -2,14 +2,14 @@ import pytest
 from django.shortcuts import resolve_url
 from rest_framework import status
 
-END_POINT = 'core:categoria-read-delete-update'
+END_POINT = 'videos:videos-read-delete-update'
 
 pytestmark = pytest.mark.django_db
 
 
-def test_read(client_auth, categoria):
+def test_read(client_auth, video):
 
-    url = resolve_url(END_POINT, categoria.id)
+    url = resolve_url(END_POINT, video.id)
 
     resp = client_auth.get(url)
 
@@ -17,9 +17,10 @@ def test_read(client_auth, categoria):
 
     body = resp.json()
 
-    assert categoria.id == body['id']
-    assert categoria.titulo == body['titulo']
-    assert categoria.cor == body['cor']
+    assert video.id == body['id']
+    assert video.titulo == body['titulo']
+    assert video.descricao == body['descricao']
+    assert video.url == body['url']
 
 
 def test_not_found(client_auth):
