+++ aluraflix_api/videos/tests/api/videos/test_videos_update.py
@@ -2,12 +2,12 @@ import pytest
 from django.shortcuts import resolve_url
 from rest_framework import status
 
-from aluraflix_api.core.tests.conftest import fake
+from aluraflix_api.conftest import fake
 
 pytestmark = pytest.mark.django_db
 
 
-END_POINT = 'core:videos-read-delete-update'
+END_POINT = 'videos:videos-read-delete-update'
 
 
 def test_full_update(client_auth, video):
